Add files via upload
A pre-trained Haar Cascade Classifier file used for detecting frontal faces in images and video frames. It is a machine learning-based approach where a set of positive and negative images are used to train the classifier. The positive images contain faces, and the negative images contain non-face objects. The algorithm uses the Haar-like features to detect the face regions in the input image or video frame. These features are simple rectangular patterns that are similar to edge detection, and they can be computed very quickly.